refactor(storage): split shared_buffer into multiple files
